pyproject.toml
[tool.poetry]
name = "Terrier"
version = "0.1.0"
description = "Classifier of Repeats"
authors = ["Robert Turnbull <robert.turnbull@unimelb.edu.au>"]
license = "Apache-2.0"
readme = "README.rst"
repository = "https://github.com/rbturnbull/terrier/"
documentation = "https://rbturnbull.github.io/terrier"
homepage = "https://github.com/rbturnbull/terrier/"
keywords = ["fastai", "torchapp", "pytorch", "deep learning", "command-line interface"]
# For classifiers see https://pypi.org/classifiers/
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Intended Audience :: Science/Research",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
torchapp = ">=0.3.7"
hierarchicalsoftmax = {git = "https://github.com/rbturnbull/hierarchicalsoftmax.git", branch="main"}
h5py = "^3.8.0"
anytree = "^2.8.0"
bio-corgi = {git = "https://github.com/rbturnbull/corgi.git", branch="hierarchical"}
pyarrow = "^12.0.0"
pyfastx = "^1.1.0"
torch = "2.0.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
ipykernel = "^6.6.1"
coverage = "^5.5"
autopep8 = "^1.5.7"
Sphinx = "^4.2.0"
nbsphinx = "^0.8.7"
sphinx-rtd-theme = "^1.0.0"
sphinx-autobuild = "^2021.3.14"
pre-commit = "^2.15.0"
sphinx-copybutton = "^0.4.0"
black = "^21.10b0"
sphinx-click = {git = "https://github.com/rbturnbull/sphinx-click.git"}

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
terrier = "terrier.apps:Terrier.main"
